Add delay effect, change GUI formatting.
Update GUI formatting to use the place geometry manager instead of the
pack geometry manager. Include a delay effect in filter_library.py
and implement the delay filter in make_gui. Note: there is an issue
with dimensionality of recorded signals: sounddevice is set up to
record to a 2d array, and my filters are set up for 1d arrays. Not
immediately clear how to resolve this without headache.